requirements.txt
tensorflow
keras
numpy
Pillow
pip
streamlit
opencv-python-headless
